Move `CargoTargetSpec` and friends to cargo_target_spec module
