+++ crates/ra_lsp_server/src/lib.rs
@@ -1,4 +1,5 @@
 mod caps;
+mod cargo_target_spec;
 mod conv;
 mod main_loop;
 mod project_model;
